refactor(serverless): centralize delivery tracking
