+++ packages/dd-trace/src/ci-visibility/exporters/agentless/request-tracker.js
@@ -30,7 +30,7 @@ class TestOptimizationRequestTracker {
    */
   flush (done, options) {
     if (options?.deadline === undefined) {
-      BaseWriter.prototype.flushDirect.call(this.#writer, done, options)
+      BaseWriter.prototype._flush.call(this.#writer, done, options)
       return
     }
 
@@ -72,7 +72,7 @@ class TestOptimizationRequestTracker {
     const previousFinalFlush = this.#activeFinalFlush
     this.#activeFinalFlush = finalFlush
     try {
-      BaseWriter.prototype.flushDirect.call(this.#writer, (error) => {
+      BaseWriter.prototype._flush.call(this.#writer, (error) => {
         finalFlush.error ||= error
         finalFlush.writerDone = true
         this.#finishFinalFlush(finalFlush)
