+++ packages/dd-trace/src/exporters/agent/index.js
@@ -3,7 +3,7 @@
 const { URL } = require('url')
 const log = require('../../log')
 const { createServerlessDeliveryTracker } = require('../../serverless')
-const { flushWriter, trackDelivery } = require('../common/flush')
+const { flushWriter, trackDelivery } = require('../../serverless/telemetry-delivery-tracker')
 const Writer = require('./writer')
 
 class AgentExporter {
