+++ packages/dd-trace/src/serverless/telemetry-delivery-tracker.js
@@ -52,4 +52,32 @@ class TelemetryDeliveryTracker {
   }
 }
 
+/**
+ * Starts a delivery through the serverless tracker when one is active.
+ * Outside supported serverless platforms, delivery keeps the existing callback path.
+ *
+ * @param {TelemetryDeliveryTracker | undefined} tracker
+ * @param {(done?: () => void) => void} deliver
+ * @param {(() => void) | undefined} done
+ */
+function trackDelivery (tracker, deliver, done) {
+  if (tracker) return tracker.track(deliver, done)
+  deliver(done)
+}
+
+/**
+ * Flushes a writer through the serverless delivery tracker.
+ * `flushDirect` avoids re-entering an automatic writer flush lifecycle hook.
+ *
+ * @param {{ flush: (done?: () => void) => void, flushDirect?: (done?: () => void) => void }} writer
+ * @param {TelemetryDeliveryTracker | undefined} tracker
+ * @param {(() => void) | undefined} done
+ */
+function flushWriter (writer, tracker, done) {
+  const flush = writer.flushDirect ?? writer.flush
+  return trackDelivery(tracker, flush.bind(writer), done)
+}
+
 module.exports = TelemetryDeliveryTracker
+module.exports.flushWriter = flushWriter
+module.exports.trackDelivery = trackDelivery
